build: The following dependencies have been added to the project:
PostgreSQL for connecting to and interacting with the database123.

Liquibase for managing database migrations, allowing control over changes to the database schema1.

SnakeYaml for reading configuration files in YAML format, which simplifies application setup1.

DotEnv-Java for managing environment variables, which allows you to securely and conveniently set configuration parameters without storing them in the code

Liquibase connection configured for database migration management. A migration script (changelog) has been prepared, containing definitions to create the necessary tables and establish relationships between them123. The script ensures automatic creation and updating of the database schema when the application is launched, which simplifies deployment and maintenance